pyproject.toml
[project]
name = "ai-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckduckgo-search>=8.0.1",
    "fastmcp>=2.2.10",
    "langchain-anthropic>=0.3.12",
    "langchain-community>=0.3.23",
    "langchain-experimental>=0.3.4",
    "langchain-mcp-adapters>=0.0.10",
    "langgraph>=0.4.2",
    "langgraph-checkpoint-sqlite>=2.0.7",
    "matplotlib>=3.10.1",
    "mcp>=1.7.1",
    "pandas>=2.2.3",
    "yfinance>=0.2.59",
]
